+++ modules/expressions/src/index.ts
@@ -5,9 +5,9 @@
 export type {
   ExpressionEvaluationOptions,
   ExpressionFunction,
-  ExpressionFunctionLibrary,
-} from "./function-libraries";
-export type { Expression, ExpressionContext } from "./expression-eval";
+  ExpressionFunctionLibrary
+} from './function-libraries';
+export type {BinaryOperator, Expression, ExpressionContext, UnaryOperator} from './expression-eval';
 export {
   addBinaryOp,
   addUnaryOp,
@@ -15,14 +15,11 @@ export {
   compileAsync,
   eval,
   evalAsync,
-  parse,
-} from "./expression-eval";
+  parse
+} from './expression-eval';
 export {
   BASIC_MATH_FUNCTION_LIBRARY,
   GEOSPATIAL_FUNCTION_LIBRARY,
-  mergeFunctionLibraries,
-} from "./function-libraries";
-export {
-  parseExpressionString,
-  type AccessorFunction,
-} from "./parse-expression-string";
+  mergeFunctionLibraries
+} from './function-libraries';
+export {parseExpressionString, type AccessorFunction} from './parse-expression-string';